fix: test confifg
