+++ backend/src/test/java/ca/bc/gov/restapi/results/common/clamav/ClamAvClientTest.java
@@ -4,7 +4,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.*;
 
 import java.time.Duration;
 import org.junit.jupiter.api.DisplayName;
